requirements.txt
appdirs==1.4.4
asgiref==3.4.1
attrs==21.2.0
cached-property==1.5.2
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
decorator==5.0.9
defusedxml==0.7.1
Django==3.2.5
django-environ==0.4.5
django-modeltranslation==0.17.3
django-rosetta==0.9.6
future==0.18.2
geocoder==1.38.1
gunicorn==20.1.0
idna==3.2
isodate==0.6.0
lxml==4.6.3
mysqlclient==2.0.3
Pillow==8.3.1
pkg_resources==0.0.0
polib==1.1.1
psycopg2==2.9.1
psycopg2-binary==2.9.1
pytz==2021.1
ratelim==0.1.6
requests==2.26.0
requests-file==1.5.1
requests-toolbelt==0.9.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
zeep==4.0.0
